// DVG.CRM.XeCung.ClientSide/dvg-crm-xecung-final-client/src/app/models/productioncost.ts
export class ProductionCost {
    constructor() {
        this.SpendDate = null;
        this.CostContent = 0;
    }
    Id: number;
    ServiceID: string;
    ProductionCostType: number;
    CostType: number;
    CostTypeText: string;
    CostContent: number;
    CostContentText: string;
    Amount: string;
    SpendDate: string;
}